Add salary account functionality, balance increment API, and validation to prevent employees from creating multiple salary accounts

+++ Backend/CachingAndConcurrentTransactionManagement/Module11/11.6/TransactionalAnnotationInSpringBoot/cachingApp/src/main/java/com/example/cachingApp/repositories/SalaryAccountRepository.java
@@ -4,6 +4,6 @@ package com.example.cachingApp.repositories;
 import com.example.cachingApp.entities.SalaryAccountEntity;
 import org.springframework.data.repository.CrudRepository;
 
-public interface SalaryAccountRepository extends CrudRepository<SalaryAccountEntity, Integer> {
+public interface SalaryAccountRepository extends CrudRepository<SalaryAccountEntity, Long> {
 
 }
